+++ Report_App_WASM/Server/Models/ReportGenerationLog.cs
@@ -2,13 +2,13 @@
 
 public class ReportGenerationLog : IExcludeAuditTrail
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public DateTime CreatedAt { get; set; }
     [MaxLength(200)] public string? CreatedBy { get; set; }
-    public int DataProviderId { get; set; }
+    public long DataProviderId { get; set; }
     [MaxLength(250)] public string? ProviderName { get; set; }
-    public int TaskLogId { get; set; }
-    public int ScheduledTaskId { get; set; }
+    public long TaskLogId { get; set; }
+    public long ScheduledTaskId { get; set; }
     [MaxLength(600)] public string? ReportName { get; set; }
     [MaxLength(600)] public string? SubName { get; set; }
     [MaxLength(60)] public string? FileType { get; set; }
