+++ Report_App_WASM/Server/Models/ScheduledTaskDistributionList.cs
@@ -2,7 +2,7 @@
 
 public class ScheduledTaskDistributionList : BaseTraceability
 {
-    public int ScheduledTaskDistributionListId { get; set; }
+    public long ScheduledTaskDistributionListId { get; set; }
     [MaxLength(4000)] public string Recipients { get; set; } = "[]";
     [MaxLength(4000)] public string? EmailMessage { get; set; }
     public virtual ScheduledTask? ScheduledTask { get; set; }
