+++ Report_App_WASM/Shared/ApiExchanges/RemoteDataPayload.cs
@@ -5,7 +5,7 @@ public class RemoteDataPayload
     public RemoteDbCommandParameters? Values { get; init; }
     public bool LogPayload { get; init; } = true;
     public bool CalculateTotalElements { get; set; }
-    public int QueryId { get; init; }
+    public long QueryId { get; init; }
     public string? QueryName { get; init; }
     public string? ProviderName { get; init; }
     public bool PivotTable { get; init; }
