+++ Report_App_WASM/Shared/DTO/TaskLogDto.cs
@@ -2,12 +2,12 @@
 
 public class TaskLogDto : IDto
 {
-    public int TaskLogId { get; set; }
-    public int ScheduledTaskId { get; set; }
+    public long TaskLogId { get; set; }
+    public long ScheduledTaskId { get; set; }
     public DateTime StartDateTime { get; set; }
     public DateTime EndDateTime { get; set; }
     public int DurationInSeconds { get; set; }
-    public int DataProviderId { get; set; }
+    public long DataProviderId { get; set; }
     [MaxLength(250)] public string? ProviderName { get; set; }
     [MaxLength(250)] public string? JobDescription { get; set; }
     [MaxLength(60)] public string? Type { get; set; }
